requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
matplotlib
pandas